separate out loading reducers from category reducers and add to constants
