+++ src/category/constants.js
@@ -3,3 +3,7 @@ export const CATEGORY_ADD = 'CATEGORY_ADD';
 export const CATEGORY_UPDATE = 'CATEGORY_UPDATE';
 export const CATEGORY_REMOVE = 'CATEGORY_REMOVE';
 export const CATEGORY_ERROR = 'CATEGORY_ERROR';
+
+export const RESPONSE_LOAD = 'RESPONSE_LOAD';
+export const PAGE_LOAD = 'PAGE_LOAD';
+export const ERROR_LOAD = 'ERROR_LOAD';
