+++ src/category/reducer.js
@@ -23,24 +23,3 @@ export function categoryReducer(state = [], { type, payload }) {  //destructured
       return state;
   }
 }
-
-export function pageLoad(state = false, { type, payload }) {
-  switch(type) {
-    default:
-      return state;
-  }
-}
-
-export function pageError(state = null, { type, payload }) {
-  switch(type) {
-    case actions.CATEGORY_LOAD:
-    case actions.CATEGORY_ADD:
-    case actions.CATEGORY_UPDATE:
-      return null;
-    case actions.CATEGORY_REMOVE:
-    case actions.CATEGORY_ERROR:
-      return payload;
-    default:
-      return state;
-  }
-}
